+++ cmd/hidden_lake/messenger/internal/utils/types.go
@@ -10,6 +10,7 @@ const (
 
 type SMessageInfo struct {
 	FFileName  string `json:"filename"` // can be ""
-	FMessage   string `json:"message"`
 	FTimestamp string `json:"timestamp"`
+	FSenderID  string `json:"sender_id"`
+	FMessage   string `json:"message"`
 }