+++ cmd/hidden_lake/messenger/pkg/settings/settings.go
@@ -21,6 +21,10 @@ const (
 )
 
 const (
+	CHeaderSenderId = "Hl-Messenger-Sender-Id"
+)
+
+const (
 	CDefaultInterfaceAddress = "127.0.0.1:9591"
 	CDefaultIncomingAddress  = "127.0.0.1:9592"
 )