+++ cmd/hidden_lake/service/pkg/request/request.go
@@ -3,6 +3,8 @@ package request
 import (
 	"fmt"
 
+	"github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
+	"github.com/number571/go-peer/pkg/crypto/random"
 	"github.com/number571/go-peer/pkg/encoding"
 )
 
@@ -19,10 +21,14 @@ type sRequest struct {
 }
 
 func NewRequest(pMethod, pHost, pPath string) IRequest {
+	prng := random.NewStdPRNG()
 	return &sRequest{
 		FMethod: pMethod,
 		FHost:   pHost,
 		FPath:   pPath,
+		FHead: map[string]string{
+			settings.CHeaderRequestId: prng.GetString(settings.CHandleRequestIDSize),
+		},
 	}
 }
 
@@ -50,8 +56,10 @@ func (p *sRequest) ToString() string {
 }
 
 func (p *sRequest) WithHead(pHead map[string]string) IRequest {
-	p.FHead = make(map[string]string)
 	for k, v := range pHead {
+		if k == settings.CHeaderRequestId {
+			panic("an attempt to overwrite the header")
+		}
 		p.FHead[k] = v
 	}
 	return p