+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Update `pkg`: append new package 'state'
 - Update `cmd/hidden_lake/service`: append option 'share' into the config
+- Update `cmd/hidden_lake/messenger`: append sender's identificators
 
 ### CHANGES